Decouple Subsystem and options bootstrapping.
Necessary for switching the relevant files to call-by-name.
