+++ src/python/pants/core/register.py
@@ -7,6 +7,7 @@ These are always activated and cannot be disabled.
 """
 
 from pants.backend.codegen import export_codegen_goal
+from pants.build_graph import build_configuration
 from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants.core.environments import rules as environments_rules
 from pants.core.environments.target_types import (
@@ -53,12 +54,14 @@ from pants.core.util_rules import (
     archive,
     config_files,
     external_tool,
+    misc,
     source_files,
     stripped_source_files,
     subprocess_environment,
     system_binaries,
 )
 from pants.core.util_rules.wrap_source import wrap_source_rule_and_target
+from pants.engine.internals import options_parsing
 from pants.engine.internals.parametrize import Parametrize
 from pants.goal import anonymous_telemetry, stats_aggregator
 from pants.source import source_root
@@ -92,10 +95,13 @@ def rules():
         *adhoc_binaries.rules(),
         *anonymous_telemetry.rules(),
         *archive.rules(),
+        *build_configuration.rules(),
         *config_files.rules(),
         *environments_rules.rules(),
         *external_tool.rules(),
         *git.rules(),
+        *misc.rules(),
+        *options_parsing.rules(),
         *source_files.rules(),
         *source_root.rules(),
         *stats_aggregator.rules(),
