+++ src/python/pants/engine/internals/options_parsing.py
@@ -3,15 +3,12 @@
 
 from dataclasses import dataclass
 
-from pants.build_graph.build_configuration import BuildConfiguration
 from pants.engine.internals.session import SessionValues
 from pants.engine.rules import collect_rules, rule
 from pants.engine.unions import UnionMembership
-from pants.option.global_options import GlobalOptions, KeepSandboxes, NamedCachesDirOption
 from pants.option.options import Options
 from pants.option.options_bootstrapper import OptionsBootstrapper
-from pants.option.scope import Scope, ScopedOptions
-from pants.util.logging import LogLevel
+from pants.option.scope import OptionsParsingSettings, Scope, ScopedOptions
 from pants.util.memo import memoized_property
 
 
@@ -24,7 +21,7 @@ class _Options:
     """
 
     options_bootstrapper: OptionsBootstrapper
-    build_config: BuildConfiguration
+    options_parsing_settings: OptionsParsingSettings
     union_membership: UnionMembership
 
     def __post_init__(self) -> None:
@@ -34,39 +31,28 @@ class _Options:
 
     @memoized_property
     def options(self) -> Options:
-        return self.options_bootstrapper.full_options(self.build_config, self.union_membership)
+        return self.options_bootstrapper.full_options(
+            self.options_parsing_settings.known_scope_infos,
+            self.union_membership,
+            self.options_parsing_settings.allow_unknown_options,
+        )
 
 
 @rule
 async def parse_options(
-    build_config: BuildConfiguration,
+    options_parsing_settings: OptionsParsingSettings,
     session_values: SessionValues,
     union_membership: UnionMembership,
 ) -> _Options:
     # TODO: Once the OptionsBootstrapper has been removed from all relevant QueryRules, this lookup
     # should be extracted into a separate @rule.
     options_bootstrapper = session_values[OptionsBootstrapper]
-    return _Options(options_bootstrapper, build_config, union_membership)
+    return _Options(options_bootstrapper, options_parsing_settings, union_membership)
 
 
 @rule
 async def scope_options(scope: Scope, options: _Options) -> ScopedOptions:
     return ScopedOptions(scope, options.options.for_scope(scope.scope))
-
-
-@rule
-async def log_level(global_options: GlobalOptions) -> LogLevel:
-    return global_options.level
-
-
-@rule
-async def extract_keep_sandboxes(global_options: GlobalOptions) -> KeepSandboxes:
-    return GlobalOptions.resolve_keep_sandboxes(global_options.options)
-
-
-@rule
-async def extract_named_caches_dir_option(global_options: GlobalOptions) -> NamedCachesDirOption:
-    return NamedCachesDirOption(global_options.named_caches_dir)
 
 
 def rules():
