+++ src/python/pants/init/options_initializer.py
@@ -26,8 +26,8 @@ from pants.init.extension_loader import (
 )
 from pants.init.plugin_resolver import PluginResolver
 from pants.init.plugin_resolver import rules as plugin_resolver_rules
+from pants.option.bootstrap_options import DynamicRemoteOptions
 from pants.option.errors import UnknownFlagsError
-from pants.option.global_options import DynamicRemoteOptions
 from pants.option.options import Options
 from pants.option.options_bootstrapper import OptionsBootstrapper
 from pants.util.requirements import parse_requirements_file
@@ -164,7 +164,9 @@ class OptionsInitializer:
         raise_: bool,
     ) -> Options:
         with self.handle_unknown_flags(options_bootstrapper, env, raise_=raise_):
-            return options_bootstrapper.full_options(build_config, union_membership)
+            return options_bootstrapper.full_options(
+                build_config.known_scope_infos, union_membership, build_config.allow_unknown_options
+            )
 
     @contextmanager
     def handle_unknown_flags(
@@ -187,7 +189,7 @@ class OptionsInitializer:
                 options_bootstrapper, args=("dummy_first_arg",)
             )
             options = no_arg_bootstrapper.full_options(
-                build_config,
+                build_config.known_scope_infos,
                 union_membership=UnionMembership.empty(),
             )
             FlagErrorHelpPrinter(options).handle_unknown_flags(err)
