+++ src/python/pants/option/subsystem.py
@@ -11,7 +11,9 @@ from collections.abc import Callable, Iterable, Sequence
 from typing import TYPE_CHECKING, Any, ClassVar, TypeVar, cast
 
 from pants.engine.env_vars import EnvironmentVars, EnvironmentVarsRequest
+from pants.engine.internals.options_parsing import scope_options
 from pants.engine.internals.selectors import AwaitableConstraints, Get
+from pants.engine.rules import implicitly
 from pants.engine.unions import UnionMembership, UnionRule, distinct_union_type_per_subclass
 from pants.option.errors import OptionsError
 from pants.option.option_types import OptionInfo, collect_options_info
@@ -51,6 +53,11 @@ class _SubsystemMeta(ABCMeta):
                     ),
                     {},
                 )
+            # A marker that allows initialization code to check for EnvironmentAware subclasses
+            # without having to import Subsystem.EnvironmentAware and use issubclass(), which
+            # can cause a dependency cycle.
+            self.EnvironmentAware.__subsystem_environment_aware__ = True
+
             self.EnvironmentAware.subsystem = self
 
 
@@ -64,6 +71,10 @@ class Subsystem(metaclass=_SubsystemMeta):
     best rendering, use soft wrapping (e.g. implicit string concatenation) within paragraphs, but
     hard wrapping (`\n`) to separate distinct paragraphs and/or lists.
     """
+
+    # A marker that allows initialization code to check for Subsystem subclasses without
+    # having to import Subsystem and use issubclass(), which can cause a dependency cycle.
+    __subsystem__ = True
 
     options_scope: str
     help: ClassVar[str | Callable[[], str]]
@@ -205,10 +216,13 @@ class Subsystem(metaclass=_SubsystemMeta):
             parameters=FrozenDict(),
             awaitables=(
                 AwaitableConstraints(
-                    rule_id=None,
+                    rule_id="pants.engine.internals.options_parsing.scope_options",
                     output_type=ScopedOptions,
-                    explicit_args_arity=0,
-                    input_types=(Scope,),
+                    explicit_args_arity=1,
+                    # NB: For a call-by-name, the input_types are the explicit ones provided
+                    # in a map passed to **implicitly(), and our call to this rule uses a
+                    # positional arg and an empty **implicitly(), so input_types are empty here.
+                    input_types=tuple(),
                     is_effect=False,
                 ),
             ),
@@ -316,7 +330,7 @@ class Subsystem(metaclass=_SubsystemMeta):
 
 
 async def _construct_subsystem(subsystem_typ: type[_SubsystemT]) -> _SubsystemT:
-    scoped_options = await Get(ScopedOptions, Scope(str(subsystem_typ.options_scope)))
+    scoped_options = await scope_options(Scope(str(subsystem_typ.options_scope)), **implicitly())
     return subsystem_typ(scoped_options.options)
 
 
@@ -334,6 +348,9 @@ async def _construct_env_aware(
     t.env_tgt = env_tgt
 
     if t.env_vars_used_by_options:
+        # TODO: Switching this to call-by-name will require some refactoring, since
+        # the environment_vars_subset() rule depends on EnvironmentSubsystem, which
+        # of course depends on Subsystem...
         t._options_env = await Get(
             EnvironmentVars, EnvironmentVarsRequest(t.env_vars_used_by_options)
         )
