+++ src/python/pants/pantsd/pants_daemon_core.py
@@ -15,7 +15,7 @@ from pants.engine.internals.native_engine import PyExecutor
 from pants.engine.unions import UnionMembership
 from pants.init.engine_initializer import EngineInitializer, GraphScheduler
 from pants.init.options_initializer import OptionsInitializer
-from pants.option.global_options import AuthPluginResult, DynamicRemoteOptions
+from pants.option.bootstrap_options import AuthPluginResult, DynamicRemoteOptions
 from pants.option.option_value_container import OptionValueContainer
 from pants.option.options_bootstrapper import OptionsBootstrapper
 from pants.option.options_diff import summarize_dynamic_options_diff, summarize_options_map_diff
